kafka: move topic utilities into server
These utilities were miscategorized as generic utilities in the previous
refactoring stage. They may eventually be refactored as generic, but
they currently have some dependencies on redpanda flavor of things from
cluster:: namespace and are only used in the server.
