+++ src/v/kafka/server/handlers/create_topics.cc
@@ -12,8 +12,8 @@
 #include "cluster/topics_frontend.h"
 #include "kafka/errors.h"
 #include "kafka/protocol/timeout.h"
-#include "kafka/protocol/topics/topic_utils.h"
-#include "kafka/protocol/topics/types.h"
+#include "kafka/server/handlers/topics/topic_utils.h"
+#include "kafka/server/handlers/topics/types.h"
 #include "kafka/types.h"
 #include "model/metadata.h"
 #include "utils/to_string.h"
