+++ src/v/kafka/server/tests/delete_topics_test.cc
@@ -11,7 +11,7 @@
 #include "kafka/protocol/delete_topics.h"
 #include "kafka/protocol/metadata.h"
 #include "kafka/protocol/schemata/create_topics_request.h"
-#include "kafka/protocol/topics/types.h"
+#include "kafka/server/handlers/topics/types.h"
 #include "kafka/types.h"
 #include "redpanda/application.h"
 #include "redpanda/tests/fixture.h"
