+++ src/v/kafka/server/tests/topic_recreate_test.cc
@@ -12,7 +12,7 @@
 #include "kafka/protocol/delete_topics.h"
 #include "kafka/protocol/metadata.h"
 #include "kafka/protocol/produce.h"
-#include "kafka/protocol/topics/types.h"
+#include "kafka/server/handlers/topics/types.h"
 #include "kafka/types.h"
 #include "model/fundamental.h"
 #include "model/record_batch_reader.h"
